requirements.txt
google.generativeai
gtts 
playsound 
SpeechRecognition
pyaudio